requirements.txt
streamlit
pandas
numpy
scikit-learn
openai>=1.0.0
python-dotenv
